feat: add cheney algorithm based gc for heap allocator - also cargo fmt & clippy on lepton_vm

+++ crates/lepton_vm/src/values.rs
@@ -16,12 +16,12 @@ pub enum Value {
     /// A simple boolean
     Bool(bool),
 
-    /// A unique create-once value which should always be unique up to MAX_u64 (which should realistically be unreachable) 
-    Tag(u64),       
+    /// A unique create-once value which should always be unique up to `MAX_u64` (which should realistically be unreachable)
+    Tag(u64),
 
     /// Integer handle pointing into the heap at an Object
-    Object(u32),    
+    Object(usize),
 
     /// Integer handle pointing into the heap at a List
-    List(u32),      
+    List(usize),
 }
